chore: refactor case_file to new location
