+++ backend/.gitignore
@@ -1,3 +1,6 @@
 node_modules
+prisma/case_management/generated
+prisma/investigation/generated
 # Keep environment variables out of version control
 .env
+
