--- backend/src/case_file/entities/prevention.entity.ts
@@ -1,5 +1,0 @@
-import { CaseFileAction } from "../../case_management/case_file_action/entities/case_file_action.entity";
-
-export class Prevention {
-  actions: CaseFileAction[];
-}
