+++ backend/src/case_management/case_file/case_file.resolver.ts
@@ -2,10 +2,10 @@ import { Resolver, Query, Mutation, Args } from "@nestjs/graphql";
 import { CaseFileService } from "./case_file.service";
 import { CreateAssessmentInput, CreateEquipmentInput, CreatePreventionInput } from "./dto/create-case_file.input";
 import { UpdateAssessmentInput, UpdateEquipmentInput, UpdatePreventionInput } from "./dto/update-case_file.input";
-import { JwtRoleGuard } from "../auth/jwtrole.guard";
+import { JwtRoleGuard } from "../../auth/jwtrole.guard";
 import { UseGuards } from "@nestjs/common";
-import { Role, coreRoles } from "../enum/role.enum";
-import { Roles } from "../auth/decorators/roles.decorator";
+import { Role, coreRoles } from "../../enum/role.enum";
+import { Roles } from "../../auth/decorators/roles.decorator";
 import { ReviewInput } from "./dto/review-input";
 import { CreateNoteInput } from "./dto/note/create-note.input";
 import { UpdateNoteInput } from "./dto/note/update-note.input";
