+++ backend/src/case_management/case_file/case_file.service.ts
@@ -1,19 +1,19 @@
 import { Injectable, Logger } from "@nestjs/common";
 import { CreateAssessmentInput, CreateCaseInput, CreatePreventionInput } from "./dto/create-case_file.input";
 import { UpdateAssessmentInput, UpdateEquipmentInput, UpdatePreventionInput } from "./dto/update-case_file.input";
-import { CaseManagementPrismaService } from "../prisma/cm/prisma.cm.service";
+import { CaseManagementPrismaService } from "../../prisma/cm/prisma.cm.service";
 import { CaseFile } from "./entities/case_file.entity";
 import { GraphQLError } from "graphql";
-import { ACTION_CODES } from "../common/action_codes";
+import { ACTION_CODES } from "../../common/action_codes";
 import { CreateNoteInput } from "./dto/note/create-note.input";
 import { UpdateNoteInput } from "./dto/note/update-note.input";
 import { DeleteNoteInput } from "./dto/note/delete-note.input";
-import { ACTION_TYPE_CODES } from "../common/action_type_codes";
+import { ACTION_TYPE_CODES } from "../../common/action_type_codes";
 import { ReviewInput } from "./dto/review-input";
-import { CaseFileActionService } from "../case_management/case_file_action/case_file_action.service";
+import { CaseFileActionService } from "../case_file_action/case_file_action.service";
 import { Equipment } from "./entities/equipment.entity";
 import { DeleteEquipmentInput } from "./dto/equipment/delete-equipment.input";
-import { action, Prisma, PrismaClient } from "../../prisma/case_management/generated";
+import { action, Prisma, PrismaClient } from "../../../prisma/case_management/generated";
 import { DefaultArgs } from "@prisma/client/runtime/library";
 import { CreateWildlifeInput } from "./dto/wildlife/create-wildlife-input";
 import { WildlifeInput } from "./dto/wildlife/wildlife-input";
