+++ backend/src/case_management/case_file/case_file_types.graphql
