+++ backend/src/case_management/case_file/dto/assessment-details.input.ts
