+++ backend/src/case_management/case_file/dto/equipment-action.ts
