+++ backend/src/case_management/case_file/dto/equipment/create-equipment-details.input.ts
