+++ backend/src/case_management/case_file/dto/equipment/delete-equipment.input.ts
