+++ backend/src/case_management/case_file/dto/equipment/update-equipment-details.input.ts
