+++ backend/src/case_management/case_file/dto/subject-query-result.ts
