+++ backend/src/case_management/case_file/dto/wildlife/delete-wildlife-input.ts
