+++ backend/src/case_management/case_file/dto/wildlife/ear-tag-input.ts
