+++ backend/src/case_management/case_file/entities/assessment.entity.ts
@@ -1,4 +1,4 @@
-import { CaseFileAction } from "../../case_management/case_file_action/entities/case_file_action.entity";
+import { CaseFileAction } from "../../case_file_action/entities/case_file_action.entity";
 import { KeyValuePair } from "./key-value-pair";
 
 export class Assessment {
