+++ backend/src/case_management/case_file/entities/authorization-outcome.entity.ts
