+++ backend/src/case_management/case_file/entities/case_file.entity.ts
