+++ backend/src/case_management/case_file/entities/note.entity.ts
@@ -1,0 +1,7 @@
+import { CaseFileAction } from "../../case_file_action/entities/case_file_action.entity";
+
+export class Note {
+  id: string;
+  note: string;
+  actions: CaseFileAction[];
+}
