+++ backend/src/case_management/case_file/entities/prevention.entity.ts
@@ -1,0 +1,5 @@
+import { CaseFileAction } from "../../case_file_action/entities/case_file_action.entity";
+
+export class Prevention {
+  actions: CaseFileAction[];
+}
